// src/services/types.ts
  export declare type Details = {
    url: string;
    name: string;
    description: string;
    imageUrl: string;
    review: Review;
    genres: string[];
    keywords: string;
    actors: Actor[];
    director: Director[];
    duration: string;
    id: string;
    rating: string;
  };
// ...
  export interface Review {
    "@type": string;
    itemReviewed: ItemReviewed;
    author: Author;
    dateCreated: string;
    inLanguage: string;
    name: string;
    reviewBody: string;
    reviewRating: ReviewRating;
  }
  
  export interface ItemReviewed {
    "@type": string;
    url: string;
  }
  
  export interface Author {
    "@type": string;
    name: string;
  }
  
  export interface ReviewRating {
    "@type": string;
    worstRating: number;
    bestRating: number;
    ratingValue: number;
  }
// ...
  export interface Actor {
    "@type": string;
    url: string;
    name: string;
  }
  
  export interface Director {
    "@type": string;
    url: string;
    name: string;
  }